Forecasting weather in 5 days,,

+++ app/src/main/java/com/robotsfoundation/weatherapp/Model/Main.java
@@ -3,7 +3,7 @@ package com.robotsfoundation.weatherapp.Model;
 public class Main {
 
     private double temp;
-    private double pressure;
+    private float pressure;
     private int humidity;
     private double temp_min;
     private double temp_max;
@@ -21,15 +21,15 @@ public class Main {
         this.temp = temp;
     }
 
-    public double getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(double pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public int  getHumidity() {
         return humidity;
     }
